libroach: initial split of db.cc
Long overdue work to start splitting up `db.cc`. The initial work moves
`mvccScanner` into `mvcc.h` and creates `status.h` to hold the *Status()
utilities and `timestamp.h` to hold the timestamp routines. Added the
`cockroach` namespace which should eventually hold of the Cockroach C++
code.

No functional change, this is all code movement.

Release note: None
