+++ c-deps/libroach/db.h
@@ -12,35 +12,63 @@
 // implied.  See the License for the specific language governing
 // permissions and limitations under the License.
 
+#pragma once
+
 #include <libroach.h>
+#include <memory>
 #include <rocksdb/comparator.h>
 #include <rocksdb/iterator.h>
 #include <rocksdb/status.h>
 #include <rocksdb/write_batch.h>
 #include <rocksdb/write_batch_base.h>
 
+struct DBIterator {
+  std::unique_ptr<rocksdb::Iterator> rep;
+  std::unique_ptr<rocksdb::WriteBatch> kvs;
+  std::unique_ptr<rocksdb::WriteBatch> intents;
+};
+
+namespace cockroach {
+
 // DBOpenHook is called at the beginning of DBOpen. It can be implemented in CCL code.
 rocksdb::Status DBOpenHook(const std::string& db_dir, const DBOptions opts);
 
 // ToDBSlice returns a DBSlice from a rocksdb::Slice
-DBSlice ToDBSlice(const rocksdb::Slice& s);
+inline DBSlice ToDBSlice(const rocksdb::Slice& s) {
+  DBSlice result;
+  result.data = const_cast<char*>(s.data());
+  result.len = s.size();
+  return result;
+}
 
-// ToString returns a c++ string with the contents of a DBSlice.
-std::string ToString(DBSlice s);
+inline DBSlice ToDBSlice(const DBString& s) {
+  DBSlice result;
+  result.data = s.data;
+  result.len = s.len;
+  return result;
+}
 
-// ToString returns a c++ string with the contents of a DBString.
-std::string ToString(DBString s);
+// ToDBString converts a rocksdb::Slice to a DBString.
+inline DBString ToDBString(const rocksdb::Slice& s) {
+  DBString result;
+  result.len = s.size();
+  result.data = static_cast<char*>(malloc(result.len));
+  memcpy(result.data, s.data(), s.size());
+  return result;
+}
+
+// ToString converts a DBSlice/DBString to a C++ string.
+inline std::string ToString(DBSlice s) { return std::string(s.data, s.len); }
+inline std::string ToString(DBString s) { return std::string(s.data, s.len); }
+
+// ToSlice converts a DBSlice/DBString to a rocksdb::Slice.
+inline rocksdb::Slice ToSlice(DBSlice s) { return rocksdb::Slice(s.data, s.len); }
+inline rocksdb::Slice ToSlice(DBString s) { return rocksdb::Slice(s.data, s.len); }
 
 // MVCC keys are encoded as <key>[<wall_time>[<logical>]]<#timestamp-bytes>. A
 // custom RocksDB comparator (DBComparator) is used to maintain the desired
 // ordering as these keys do not sort lexicographically correctly.
 std::string EncodeKey(DBKey k);
-
-// ToDBStatus converts a rocksdb Status to a DBStatus.
-DBStatus ToDBStatus(const rocksdb::Status& status);
-
-// FmtStatus formats the given arguments printf-style into a DBStatus.
-DBStatus FmtStatus(const char* fmt, ...);
 
 // CockroachComparator returns CockroachDB's custom mvcc-aware RocksDB
 // comparator. The caller does not assume ownership.
@@ -54,3 +82,5 @@ const ::rocksdb::Comparator* CockroachComparator();
 // Stats are only computed for keys between the given range.
 MVCCStatsResult MVCCComputeStatsInternal(::rocksdb::Iterator* const iter_rep, DBKey start,
                                          DBKey end, int64_t now_nanos);
+
+}  // namespace cockroach
