+++ c-deps/libroach/encoding.cc
@@ -13,7 +13,10 @@
 // permissions and limitations under the License.
 
 #include "encoding.h"
-#include "rocksdb/slice.h"
+#include <rocksdb/slice.h>
+#include "db.h"
+
+namespace cockroach {
 
 void EncodeUint32(std::string* buf, uint32_t v) {
   const uint8_t tmp[sizeof(v)] = {
@@ -56,3 +59,121 @@ bool DecodeUint64(rocksdb::Slice* buf, uint64_t* value) {
   buf->remove_prefix(N);
   return true;
 }
+
+void EncodeTimestamp(std::string& s, int64_t wall_time, int32_t logical) {
+  EncodeUint64(&s, uint64_t(wall_time));
+  if (logical != 0) {
+    EncodeUint32(&s, uint32_t(logical));
+  }
+}
+
+std::string EncodeTimestamp(DBTimestamp ts) {
+  std::string s;
+  s.reserve(kMVCCVersionTimestampSize);
+  EncodeTimestamp(s, ts.wall_time, ts.logical);
+  return s;
+}
+
+// MVCC keys are encoded as <key>[<wall_time>[<logical>]]<#timestamp-bytes>. A
+// custom RocksDB comparator (DBComparator) is used to maintain the desired
+// ordering as these keys do not sort lexicographically correctly.
+std::string EncodeKey(const rocksdb::Slice& key, int64_t wall_time, int32_t logical) {
+  std::string s;
+  const bool ts = wall_time != 0 || logical != 0;
+  s.reserve(key.size() + 1 + (ts ? 1 + kMVCCVersionTimestampSize : 0));
+  s.append(key.data(), key.size());
+  if (ts) {
+    // Add a NUL prefix to the timestamp data. See DBPrefixExtractor.Transform
+    // for more details.
+    s.push_back(0);
+    EncodeTimestamp(s, wall_time, logical);
+  }
+  s.push_back(char(s.size() - key.size()));
+  return s;
+}
+
+// MVCC keys are encoded as <key>[<wall_time>[<logical>]]<#timestamp-bytes>. A
+// custom RocksDB comparator (DBComparator) is used to maintain the desired
+// ordering as these keys do not sort lexicographically correctly.
+std::string EncodeKey(DBKey k) { return EncodeKey(ToSlice(k.key), k.wall_time, k.logical); }
+
+WARN_UNUSED_RESULT bool SplitKey(rocksdb::Slice buf, rocksdb::Slice* key,
+                                 rocksdb::Slice* timestamp) {
+  if (buf.empty()) {
+    return false;
+  }
+  const char ts_size = buf[buf.size() - 1];
+  if (ts_size >= buf.size()) {
+    return false;
+  }
+  *key = rocksdb::Slice(buf.data(), buf.size() - ts_size - 1);
+  *timestamp = rocksdb::Slice(key->data() + key->size(), ts_size);
+  return true;
+}
+
+WARN_UNUSED_RESULT bool DecodeTimestamp(rocksdb::Slice* timestamp, int64_t* wall_time,
+                                        int32_t* logical) {
+  uint64_t w;
+  if (!DecodeUint64(timestamp, &w)) {
+    return false;
+  }
+  *wall_time = int64_t(w);
+  *logical = 0;
+  if (timestamp->size() > 0) {
+    // TODO(peter): Use varint decoding here.
+    uint32_t l;
+    if (!DecodeUint32(timestamp, &l)) {
+      return false;
+    }
+    *logical = int32_t(l);
+  }
+  return true;
+}
+
+WARN_UNUSED_RESULT bool DecodeTimestamp(rocksdb::Slice buf,
+                                        cockroach::util::hlc::Timestamp* timestamp) {
+  int64_t wall_time;
+  int32_t logical;
+  if (!DecodeTimestamp(&buf, &wall_time, &logical)) {
+    return false;
+  }
+  timestamp->set_wall_time(wall_time);
+  timestamp->set_logical(logical);
+  return true;
+}
+
+WARN_UNUSED_RESULT bool DecodeKey(rocksdb::Slice buf, rocksdb::Slice* key, int64_t* wall_time,
+                                  int32_t* logical) {
+  key->clear();
+
+  rocksdb::Slice timestamp;
+  if (!SplitKey(buf, key, &timestamp)) {
+    return false;
+  }
+  if (timestamp.size() > 0) {
+    timestamp.remove_prefix(1);  // The NUL prefix.
+    if (!DecodeTimestamp(&timestamp, wall_time, logical)) {
+      return false;
+    }
+  }
+  return timestamp.empty();
+}
+
+rocksdb::Slice KeyPrefix(const rocksdb::Slice& src) {
+  rocksdb::Slice key;
+  rocksdb::Slice ts;
+  if (!SplitKey(src, &key, &ts)) {
+    return src;
+  }
+  // RocksDB requires that keys generated via Transform be comparable with
+  // normal encoded MVCC keys. Encoded MVCC keys have a suffix indicating the
+  // number of bytes of timestamp data. MVCC keys without a timestamp have a
+  // suffix of 0. We're careful in EncodeKey to make sure that the user-key
+  // always has a trailing 0. If there is no timestamp this falls out
+  // naturally. If there is a timestamp we prepend a 0 to the encoded
+  // timestamp data.
+  assert(src.size() > key.size() && src[key.size()] == 0);
+  return rocksdb::Slice(key.data(), key.size() + 1);
+}
+
+}  // namespace cockroach
